Moving `R4DataProviderFactory`
Moving `R4DataProviderFactory` to the `com.ibm.cohort.engine.data` package because it has no direct relation to measure evaluation.
